+++ core/src/core-plugins.ts
@@ -245,7 +245,7 @@ export interface HttpOptions {
    * This is used if we've had to convert the data from a JS type that needs
    * special handling in the native layer
    */
-  dataType?: 'file' | 'formData';
+  dataType?: 'file' | 'formData' | 'binary';
 }
 
 export interface HttpParams {
@@ -262,11 +262,11 @@ export interface HttpHeaders {
   [key: string]: string;
 }
 
-export interface HttpResponse {
+export interface HttpResponse<T = any> {
   /**
    * Additional data received with the Http response.
    */
-  data: any;
+  data: T;
   /**
    * The status code received from the Http response.
    */
@@ -318,7 +318,7 @@ const normalizeHttpHeaders = (headers: HttpHeaders = {}): HttpHeaders => {
  * @param params A map of url parameters
  * @param shouldEncode true if you should encodeURIComponent() the values (true by default)
  */
-const buildUrlParams = (params?: HttpParams, shouldEncode = true): string | null => {
+export const buildUrlParams = (params?: HttpParams, shouldEncode = true): string | null => {
   if (!params) return null;
 
   const output = Object.entries(params).reduce((accumulator, entry) => {
@@ -333,7 +333,7 @@ const buildUrlParams = (params?: HttpParams, shouldEncode = true): string | null
         item += `${key}=${encodedValue}&`;
       });
       // last character will always be "&" so slice it off
-      item.slice(0, -1);
+      item = item.slice(0, -1);
     } else {
       encodedValue = shouldEncode ? encodeURIComponent(value) : value;
       item = `${key}=${encodedValue}`;
@@ -343,7 +343,7 @@ const buildUrlParams = (params?: HttpParams, shouldEncode = true): string | null
   }, '');
 
   // Remove initial "&" from the reduce
-  return output.substr(1);
+  return output.substring(1);
 };
 
 /**